requirements.txt
h5py==3.14.0
matplotlib==3.10.6
numpy==2.3.3
PyYAML==6.0.2
PyYAML==6.0.3
seaborn==0.13.2
torch==2.7.1
tqdm==4.67.1
